Enable close button on prefrences window
Setting the decoration layout with the close string allowed the close button to appear, but in the top left. This is a bit different from the established theme. I'm curious if this would conflict with other windowing systems since the problem doesn't seem to exist for all users right now. Would be nice to test.